backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
paho-mqtt==2.1.0
python-dotenv==1.0.1
